+++ package.json
@@ -16,9 +16,10 @@
     "jsx"
   ],
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.4.1",
     "eslint": "^7.7.0",
-    "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-import": "^2.20.2",
+    "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint-plugin-import": "^2.22.1",
     "lerna": "^3.19.0"
   }
 }
