+++ packages/babel-plugin-jsx/src/index.ts
@@ -6,30 +6,9 @@ import { addNamed, isModule, addNamespace } from '@babel/helper-module-imports';
 import { NodePath } from '@babel/traverse';
 import transformVueJSX from './transform-vue-jsx';
 import sugarFragment from './sugar-fragment';
+import type { VueJSXPluginOptions, State } from './interface';
 
-export type State = {
-  get: (name: string) => any;
-  set: (name: string, value: any) => any;
-  opts: VueJSXPluginOptions;
-  file: BabelCore.BabelFile
-}
-
-export interface VueJSXPluginOptions {
-  /** transform `on: { click: xx }` to `onClick: xxx` */
-  transformOn?: boolean;
-  /** enable optimization or not. */
-  optimize?: boolean;
-  /** merge static and dynamic class / style attributes / onXXX handlers */
-  mergeProps?: boolean;
-  /** configuring custom elements */
-  isCustomElement?: (tag: string) => boolean;
-  /** enable object slots syntax */
-  enableObjectSlots?: boolean;
-  /** Replace the function used when compiling JSX expressions */
-  pragma?: string;
-}
-
-export type ExcludesBoolean = <T>(x: T | false | true) => x is T;
+export { VueJSXPluginOptions };
 
 const hasJSX = (parentPath: NodePath<t.Program>) => {
   let fileHasJSX = false;
