+++ packages/babel-plugin-jsx/tsconfig.json
@@ -4,7 +4,6 @@
     "rootDirs": ["./src"],
     "outDir": "dist",
     "downlevelIteration": true,
-    "types": ["node", "jest"],
     "declaration": true
   },
   "include": ["src/**/*", "global.d.ts"],
