+++ tsconfig.json
@@ -12,10 +12,15 @@
     "esModuleInterop": true,
     "removeComments": false,
     "jsx": "preserve",
-    "lib": ["esnext", "dom"],
-    "types": ["node"]
+    "lib": [
+      "esnext",
+      "dom"
+    ],
+    "types": ["node", "jest"],
   },
   "include": [
-    "global.d.ts"
+    "global.d.ts",
+    "packages/*/src",
+    "packages/*/test",
   ]
 }
